DELIA-43983: setCachedValue, getCachedValue - corrected failure response status.

+++ SystemServices/SystemServices.cpp
@@ -1649,18 +1649,18 @@ namespace WPEFramework {
          */
         uint32_t SystemServices::getCachedValue(const JsonObject& parameters,
 			JsonObject& response)
-        {
-            bool retStat = false;
-            std::string key = parameters["key"].String();
-            LOGWARN("key: %s\n", key.c_str());
-            if (key.length()) {
-                response[(key.c_str())] = m_cacheService.getValue(key).String();
-                retStat = true;
-            } else {
-                populateResponseWithError(SysSrv_UnSupportedFormat, response);
-            }
-            returnResponse(retStat);
-        }
+	{
+		bool retStat = false;
+		std::string key = parameters["key"].String();
+		LOGWARN("key: %s\n", key.c_str());
+		if (key.length() && !key.is_empty()) {
+			response[(key.c_str())] = m_cacheService.getValue(key).String();
+			retStat = true;
+		} else {
+			populateResponseWithError(SysSrv_UnSupportedFormat, response);
+		}
+		returnResponse(retStat);
+	}
 
         /***
          * @brief : To set cache value.
@@ -1675,7 +1675,7 @@ namespace WPEFramework {
             std::string key = parameters["key"].String();
             std::string value = parameters["value"].String();
 
-	    if (key.length() && value.length()) {
+	    if (key.length() && value.length() && !key.is_empty() && !value.is_empty()) {
 		    if (m_cacheService.setValue(key, value)) {
 			    retStat = true;
 		    } else {
@@ -1699,7 +1699,7 @@ namespace WPEFramework {
         {
 		bool retStat = false;
 		std::string key = parameters["key"].String();
-		if (key.length()) {
+		if (key.length() && !key.is_empty()) {
 			if (m_cacheService.contains(key)) {
 				retStat = true;
 			} else {
@@ -1721,19 +1721,19 @@ namespace WPEFramework {
         uint32_t SystemServices::removeCacheKey(const JsonObject& parameters,
                 JsonObject& response)
         {
-            bool retStat = false;
-            std::string key = parameters["key"].String();
-            if (key.length()) {
-                if (m_cacheService.remove(key)) {
-                    retStat = true;
-                } else {
-                    LOGERR("Accessing m_cacheService.remove failed\n.");
-                    populateResponseWithError(SysSrv_Unexpected, response);
-                }
-            } else {
-                populateResponseWithError(SysSrv_UnSupportedFormat, response);
-            }
-            returnResponse(retStat);
+		bool retStat = false;
+		std::string key = parameters["key"].String();
+		if (key.length() && !key.is_empty()) {
+			if (m_cacheService.remove(key)) {
+				retStat = true;
+			} else {
+				LOGERR("Accessing m_cacheService.remove failed\n.");
+				populateResponseWithError(SysSrv_Unexpected, response);
+			}
+		} else {
+			populateResponseWithError(SysSrv_UnSupportedFormat, response);
+		}
+		returnResponse(retStat);
         }
 
         /***
